Add files via upload
(v1.7) GZ files now supported. Added the recognition of a line in the batch list starting with the word - stop, which the program will stop checking at. NOTE - This means the ability to have only a portion of the list processed, stopping at a pre-determined location or point. This is a handy option with a large list or before large files that take a long time to process. It essentially means you can load a list, and process it in segments. IMPORTANT - To use multiple 'stop' lines in the list, then each must be unique, like having a trailing number or any text really (i.e. stop 1 or stop 2 or stop here etc). The LIST button on the dropbox has an extra right-click option to get the program to create a 'stop' entry for you, with a trailing timestamp to make it unique.